added the slot name and type of question paper in the academic help section
