+++ frontend/scss/_academic_help.scss
@@ -1,0 +1,76 @@
+// ************************************
+// academic help
+// ************************************
+
+.test-paper {
+  &-section {
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    gap: 1.4rem;
+    padding: 1.4rem;
+    height: 90vh;
+    overflow: scroll;
+  }
+
+  &-box {
+    position: relative;
+    background-color: var(--color-black-3);
+    border-radius: 1.4rem;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-direction: column;
+    height: 32rem;
+    overflow: hidden;
+
+    &:hover {
+      cursor: pointer;
+    }
+
+    &-img {
+      height: 24.2rem;
+    }
+
+    &-text {
+      width: 100%;
+      padding: 1.4rem;
+      display: flex;
+      align-items: center;
+      justify-content: space-between;
+      flex-direction: column;
+      background-color: var(--color-white);
+      color: var(--color-black-3);
+
+      &-btn {
+        padding: 0.6rem;
+        border-radius: 0.4rem;
+      }
+    }
+
+    &-slot {
+      padding: 0.2rem 1.2rem;
+      border-radius: 1.4rem;
+      position: absolute;
+      top: 15;
+      right: 15;
+      background-image: linear-gradient(
+        60deg,
+        var(--color-blue-1),
+        var(--color-blue-2)
+      );
+    }
+
+    &-type {
+      padding: 0.2rem 1.2rem;
+      border-radius: 1.4rem;
+      position: absolute;
+      top: 15;
+      left: 15;
+      background-image: linear-gradient(
+        60deg,
+        var(--color-blue-1),
+        var(--color-blue-2)
+      );
+    }
+  }
+}
