+++ frontend/scss/_cta_section.scss
@@ -1,0 +1,79 @@
+// ************************************
+// search form
+// ************************************
+
+.cta-field {
+  display: grid;
+  grid-template-columns: 1fr 0.2fr;
+  align-items: center;
+  justify-items: center;
+}
+
+.search {
+  &-form {
+    display: flex;
+    align-items: center;
+    gap: 1.4rem;
+    padding: 1.4rem;
+    width: 100%;
+    position: relative;
+  }
+
+  &-input {
+    width: 100%;
+    padding: 1.6rem 2.4rem;
+    font-size: 1.6rem;
+    border-radius: 3.6rem;
+    border: none;
+
+    &::placeholder {
+      font-family: inherit;
+      color: var(--color-black-1);
+    }
+
+    &:focus {
+      outline-color: var(--color-black-1);
+    }
+  }
+
+  &-btn {
+    font-size: 1.6rem;
+    padding: 1rem;
+    border-radius: 50%;
+    background-image: linear-gradient(
+      60deg,
+      var(--color-blue-1),
+      var(--color-blue-2)
+    );
+    color: var(--color-white);
+    border: none;
+    position: absolute;
+    top: 50%;
+    right: 5;
+    transform: translate(-50%, -50%);
+    box-shadow: none !important;
+  }
+}
+
+.post {
+  z-index: 99;
+
+  &-img {
+    font-size: 1.8rem;
+  }
+
+  &-btn {
+    color: var(--color-white);
+    padding: 1.4rem 2.4rem;
+    border-radius: 12rem;
+    display: flex;
+    align-items: center;
+    gap: 0.8rem;
+    box-shadow: none !important;
+    background-image: linear-gradient(
+      60deg,
+      var(--color-blue-1),
+      var(--color-blue-2)
+    );
+  }
+}
