+++ frontend/scss/_general.scss
@@ -1,0 +1,42 @@
+/* SCALE */
+/* 7.4  6.4  5.4  4.4  3.4  2.4  2re  1.8  1.6  1.4  1.3  1.0  0.8  0.6  0.4  0.2   */
+
+:root {
+  --color-black-1: #404040;
+  --color-black-2: #171717;
+  --color-black-3: #0a0a0a;
+
+  --color-white: #fff;
+
+  --color-blue-1: #439ddf;
+  --color-blue-2: #4f87ed;
+}
+
+::-webkit-scrollbar {
+  width: 0;
+}
+
+*,
+*:after,
+*:before {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+html {
+  font-size: 62.5%;
+  scroll-behavior: smooth;
+  overflow: hidden;
+}
+
+body {
+  font-family: "Nanum Gothic Coding", monospace;
+  font-size: 1.6rem;
+  font-weight: 400;
+  line-height: 1;
+  overflow: hidden;
+  scroll-behavior: smooth;
+  background-color: var(--color-black-1);
+  color: var(--color-white);
+}
