+++ frontend/scss/main.scss
@@ -1,0 +1,5 @@
+@import "./general";
+@import "./utility";
+@import "./cta_section";
+@import "./home_tab";
+@import "./academic_help";
